Fixed hook_id declarations to be gulong instead of guint. Made statusbar's hook_id static. Rework hook_id magic value, now 0 instead of -1, and use a define (HOOK_NONE) everywhere instead of 0. Any hook_id must be initialized to HOOK_NONE. Hook_id that get invalidated or reset get the same HOOK_NONE value.

+++ src/gtk/logwindow.h
@@ -44,7 +44,7 @@ struct _LogWindow
 
 	gboolean clip;
 	guint	 clip_length;
-	guint 	 hook_id;
+	gulong 	 hook_id;
 	GtkTextBuffer *buffer;
 	GtkTextTag *error_tag;
 	GtkTextMark *end_mark;
